Create README.md
##A Tribute To Paul Trouble Anderson##
This is a project for a pre course challenge set and supported by School of Code. 
The challenge was to host a web page tribute to anybody of our choosing.
I chose the DJ Paul Trouble Anderson as he was a big inspiration to me as a DJ and is often, in my opinion, undervalued for his contribution to dance music.

This was the first time I have used any of the tools *VS Code, Git/Git Hub* or any coding languages at all. 
I enjoyed the challenge but feel slightly overwhelmed at this stage with all the new knowledge I am having to make sense of. 
Hoping that with some hard work and dedication, it will all fall into place soon!